+++ Assets/PROJECT/Scripts/Characters/Lilian/Passive/LilianPassiveSO.cs
@@ -19,7 +19,7 @@ public class LilianPassiveSO : PassiveSO {
 
     public float WrathStunDuration;
     public float WrathDamage;
-    public float MaxAmountOfCorruption;
+    public int MaxAmountOfCorruption;
 
     public GameObject LilianUI;
 }
